Moving datetime related helpers into their own file.
