+++ resources/assets/components/utilities/ScholarshipCard/ScholarshipCard.js
@@ -3,7 +3,7 @@ import gql from 'graphql-tag';
 import { propType } from 'graphql-anywhere';
 
 import GalleryBlockSignup from './GalleryBlockSignup';
-import { getHumanFriendlyDate } from '../../../helpers';
+import { getHumanFriendlyDate } from '../../../helpers/datetime';
 import {
   contentfulImageSrcset,
   contentfulImageUrl,
